requirements.txt
cloudflare==2.1.0
dnspython==1.16.0